tools/opensnoop: perform name filter in kernel space
I did this while planning to adapt opensnoop further into an internal
daemon that runs for long periods of time across many thousands of
boxes, and I figured it'd be best to be as selective as possible as
early as possible in the data flow to minimize resource consumption.

However, we're pausing that work now, so this is just a speculative
optimization detached from any production rationale. Caveat Emptor.